add catch events

+++ src/main/java/simulation/SimulationClock.java
@@ -5,17 +5,17 @@ import java.util.logging.Logger;
 public class SimulationClock {
 
 	private final static Logger LOGGER = Logger.getLogger("ENSO-APP");
-	private static int simulationTime = 0;
+	private static long simulationTime = 0;
 	private final Object MUTEXTIME = new Object();
 
 	public SimulationClock() {
 	}
 
-	public int getCurrentTime() {
+	public long getCurrentTime() {
 		return simulationTime;
 	}
 
-	public void setCurrentTime(int simulationTime) {
+	public void setCurrentTime(long simulationTime) {
 		synchronized (MUTEXTIME) {
 			this.simulationTime = simulationTime;
 		}
