+++ src/main/java/simulation/SimulationStartEvent.java
@@ -2,7 +2,7 @@ package simulation;
 
 public class SimulationStartEvent extends SimulationEvent{
 	
-	public SimulationStartEvent(String name, int endTime) {
+	public SimulationStartEvent(String name, long endTime) {
 		this.name = name;
 		this.endTime = endTime;
 	}
